Instagram likes bot
This bot is based on a tutorial made by Markets and Data, which I then had to edit in order to run on my machine because it had a bunch of errors. I will be adjusting this program in the future and making an instagram auto-commenter, auto-follower, and similar tools for Youtube, Facebook, Twitter, Soundcloud etc.